fix popup, pages
