+++ src/components/Api.js
@@ -1,11 +1,11 @@
 //авторизация на сервере
-export const currentUser = {
-  currentUrl: 'https://nomoreparties.co/v1/plus-cohort-18',
-  headers: {
-    authorization: '5316090b-29fc-4b6a-8b2f-268d3472034e',
-    'Content-Type': 'application/json'
-  }
-}
+// export const currentUser = {
+//   currentUrl: 'https://nomoreparties.co/v1/plus-cohort-18',
+//   headers: {
+//     authorization: '5316090b-29fc-4b6a-8b2f-268d3472034e',
+//     'Content-Type': 'application/json'
+//   }
+// }
 
 export class Api {
   constructor({ currentUrl, headers }) {
