+++ src/utils/utils.js
@@ -1,0 +1,14 @@
+export const currentUser = {
+  currentUrl: 'https://nomoreparties.co/v1/plus-cohort-18',
+  headers: {
+    authorization: '5316090b-29fc-4b6a-8b2f-268d3472034e',
+    'Content-Type': 'application/json'
+  }
+}
+
+export const settings = {
+  inputSelector: ".popup__form-text",
+  inputErrorClass: "popup__form-text_type_error",
+  submitButtonSelector: ".popup__button",
+  inactiveButtonClass: "popup__button_disabled",
+}
