+++ webpack.config.js
@@ -4,7 +4,7 @@ const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 
 module.exports = {
-  entry: { main: './src/components/index.js' },
+  entry: { main: './src/pages/index.js' },
   output: {
     path: path.join(__dirname, 'dist'),
     filename: 'main.js',
